requirements.txt
fastapi
uvicorn[standard]
datasets
anthropic
python-dotenv
